% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captain_maps.R, R/captain_maps2.R
\name{captain_maps}
\alias{captain_maps}
\title{captain_maps}
\usage{
captain_maps(output.directory = NULL)

captain_maps(output.directory = NULL)
}
\arguments{
\item{output.directory}{Output to selected output directory}

\item{releases.data.path}{Get release data from specified path (gets recaptures from Oracle).}
}
\description{
creates maps for individual captains who participated in tag releases showing their release locations and locations of all tags recaptured from those releases.

Outputs both kml and pdf maps.

creates maps for individual captains who participated in tag releases showing their release locations and locations of all tags recaptured from those releases.

Outputs both kml and pdf maps.
}
\author{
Geraint Element
}
