% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captain_maps2.R
\name{captain_maps2}
\alias{captain_maps2}
\title{captain_maps2}
\usage{
captain_maps2(output.directory = NULL)
}
\arguments{
\item{output.directory}{Output to selected output directory}
}
\description{
creates maps for individual captains who participated in tag releases showing their release locations and locations of all tags recaptured from those releases.

Outputs both kml and pdf maps.
}
\author{
Geraint Element
}
