% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagMail.r
\name{do_not_send_email}
\alias{do_not_send_email}
\title{do_not_send_email}
\usage{
do_not_send_email(
  working_email_dir =
    "R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/Projects and Programs/Tagging/LobTag_outputs/Temp Files/Email Attachments/",
  r_drive_not_sent_email_dir =
    "R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/Projects and Programs/Tagging/LobTag_outputs/Not Sent Documents/"
)
}
\description{
moves pdf for further review, indicates human action required
}
