% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagMail.r
\name{email_rewards2}
\alias{email_rewards2}
\title{email_rewards2}
\usage{
email_rewards2(
  to = "lobtags@gmail.com",
  subject = "default subject",
  body = "default body",
  attachments = ""
)
}
\arguments{
\item{to}{email address as text}

\item{subject}{Text for subject}

\item{body}{Text for body}

\item{attachments}{path to file to attach to email}
}
\description{
Sends an email
}
