% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCrewards.R
\name{generate.reward.data}
\alias{generate.reward.data}
\title{generate.reward.data}
\usage{
generate.reward.data(region = "ScotianShelf")
}
\description{
Gather Reward data in preparation for document creation with rewards.knit.document() .
}
