% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobTagging.r
\name{get.pathdata.tid}
\alias{get.pathdata.tid}
\title{get.pathdata.tid}
\usage{
get.pathdata.tid(region = "ScotianShelf", tid = "")
}
\value{
dataframe
}
\description{
Return calculated paths by supplied TID
}
