% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCrewards.R
\name{getBetween}
\alias{getBetween}
\title{getBetween}
\usage{
getBetween(text = "", pattern = "")
}
\arguments{
\item{text}{The text feild to search}

\item{pattern}{The pattern to match}
}
\description{
Return from a text between search pattern. Helps facilitate letter format changes.
}
