\name{LobTag-package}
\alias{LobTag-package}
\alias{LobTag}
\docType{package}
\title{
\packageTitle{LobTag}
}
\description{
\packageDescription{LobTag}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LobTag}
\packageIndices{LobTag}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LobTag}

Maintainer: \packageMaintainer{LobTag}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
