% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagMail.r
\name{prep_email_send}
\alias{prep_email_send}
\title{prep_email_send}
\usage{
prep_email_send(
  r_drive_email_dir =
    "R:/Science/Population Ecology Division/Shared/!PED_Unit17_Lobster/Lobster Unit/Projects and Programs/Tagging/LobTag_outputs/Temp Files/Email Attachments/"
)
}
\description{
preps and returns info about the email that's about to be sent.
}
