% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCrewards.R
\name{rewards.knit.document}
\alias{rewards.knit.document}
\title{rewards.knit.document}
\usage{
rewards.knit.document(region = "ScotianShelf")
}
\description{
The main function call to create reward pdf from R.
}
