% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobRewards.r
\name{rewards.loop.function}
\alias{rewards.loop.function}
\title{rewards.loop.function}
\usage{
rewards.loop.function()
}
\description{
function that will generate all the rewards emails and put them in a folder
}
