% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobTagging.r
\name{shortestpaths.SC}
\alias{shortestpaths.SC}
\title{Shortestpaths.SC}
\usage{
shortestpaths.SC(
  neighborhood = 16,
  type = "least.cost",
  redo = F,
  region = "ScotianShelf",
  raster.path = "C:/bio.data/bio.lobster/data/tagging/depthraster2.tif"
)
}
\arguments{
\item{neighborhood}{The number of adjacent cell over which to calculate, defaults to 16.}

\item{type}{The type of calculation, either 'random.walk' or 'least.cost'. Defaults to 'least.cost'}

\item{redo}{Set redo = TRUE if you want to rewrite all data, FALSE to only update new entries}

\item{region}{Either 'ScotianShelf' or 'Gulf'}
}
\value{
dataframe
}
\description{
Creates and writes the shortest paths to database
}
