% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging_mapping_prep.R
\name{tagging_mapping_prep}
\alias{tagging_mapping_prep}
\title{tagging_mapping_prep}
\usage{
tagging_mapping_prep()
}
\description{
prepares all tagging data and produces useful variables, including NS basemap, for mapping tagging data

Outputs a list of mapping variables
}
\author{
Geraint Element
}
