% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagWebGUI.r
\name{upload_from_file3}
\alias{upload_from_file3}
\title{upload_from_file3}
\usage{
upload_from_file3(myfile)
}
\value{
note to use will be read in web interface
}
\description{
parses data, error check and sends to sample_ent function
}
