% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captain_maps.R
\name{captain_maps}
\alias{captain_maps}
\title{captain_maps}
\usage{
captain_maps(releases.data.path = NULL, output.directory = NULL)
}
\arguments{
\item{releases.data.path}{Get release data from specified path (gets recaptures from Oracle).}

\item{output.directory}{Output to selected output directory}
}
\description{
creates maps for individual captains who participated in tag releases showing their release locations and locations of all tags recaptured from those relesaes.

Outputs both kml and pdf maps.
}
\author{
Geraint Element
}
