% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobRewards.r
\name{rewards.chart}
\alias{rewards.chart}
\title{rewards.chart}
\usage{
rewards.chart(name = "", data = NULL, keep_anon = FALSE)
}
\arguments{
\item{name}{The name of the person}

\item{data}{The data to chart}
}
\description{
Create Reward charts.
}
