% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagMail.r
\name{send_single_email}
\alias{send_single_email}
\title{send_single_email}
\usage{
send_single_email()
}
\description{
uses latest pdf, looks up email address associated with fisher and send attachment to them.
}
