% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging_mapping_prep.R
\name{tagging_mapping_prep}
\alias{tagging_mapping_prep}
\title{tagging_mapping_prep}
\usage{
tagging_mapping_prep()
}
\arguments{
\item{releases.data.path}{Get release data from specified path (gets recaptures from Oracle).}
}
\description{
preps tagging data for mapping analysis (sf, raster, satellite basemaps, etc)

Outputs various sf objects and basemaps to begin mapping.
}
\author{
Geraint Element
}
